modificando la estructura
