--- Hola Mundo.py
@@ -1,6 +1,0 @@
-uno = 455
-print (uno)
-uno = 'Hola mundo'
-print (uno)
-print (uno)
-print (uno)
