+++ src/Hola Mundo.py
@@ -1,0 +1,6 @@
+uno = 455
+print (uno)
+uno = 'Hola mundo'
+print (uno)
+print (uno)
+print (uno)
